+++ src/gpu/graphite/render/BitmapTextRenderStep.h
@@ -31,7 +31,7 @@ public:
     const char* fragmentColorSkSL() const override;
     const char* fragmentCoverageSkSL() const override;
 
-    void writeVertices(DrawWriter*, const DrawParams&, skvx::ushort2 ssboIndices) const override;
+    void writeVertices(DrawWriter*, const DrawParams&, skvx::uint2 ssboIndices) const override;
     void writeUniformsAndTextures(const DrawParams&, PipelineDataGatherer*) const override;
 
 private:
